Use isinstance instead of type()
Using type() checks that the class of an argument is EXACTLY a given
class, and will not match if it's a subclass.

I'm using a project that uses freezegun for testing, so our date objects
are actually subclasses of `datetime.date`, and not instances of EXACTLY
that class.

This change does not break existing functionality, and should not break
even under very odd scenarios which haven't been considered so far, but
avoid unexpected failures when subclassing `datetime.date` or
`datetime.datetime`.